requirements.txt
Flask~=2.0.1
msal>=1.7,<2
requests
python-dotenv~=0.19.0
munch~=2.5.0